move mock_solver to pypsa_helpers
