+++ workflow/scripts/_pypsa_helpers.py
@@ -308,6 +308,19 @@ def load_network_for_plots(
     return n
 
 
+def mock_solve(n: pypsa.Network) -> pypsa.Network:
+    """Mock the solving step for tests
+
+    Args:
+        n (pypsa.Network): the network object
+    """
+    for c in n.iterate_components(components=["Generator", "Link", "Store", "LineType"]):
+        opt_cols = [col for col in c.df.columns if col.endswith("opt")]
+        base_cols = [col.split("_opt")[0] for col in opt_cols]
+        c.df[opt_cols] = c.df[base_cols]
+    return n
+
+
 def make_periodic_snapshots(
     year: int,
     freq: int,
